Add testFindWard test
